Refactoring nn_lib.utils into multiple files (but backwards compatible)
